corrects naming convention
