+++ lib/version.rb
@@ -1,0 +1,3 @@
+class Otask
+  VERSION = '0.2.2'
+end
